Move 1Inch api from solver crate to shared
