+++ crates/shared/src/oneinch_api.rs
@@ -3,14 +3,12 @@
 //! For more information on the HTTP API, consult:
 //! <https://docs.1inch.io/api/quote-swap>
 //! <https://api.1inch.exchange/swagger/ethereum/>
-use crate::solver::single_order_solver::SettlementError;
-use anyhow::{anyhow, ensure, Context, Result};
-use derive_more::From;
+use crate::solver_utils::{deserialize_prefixed_hex, Slippage};
+use anyhow::{ensure, Context, Result};
 use ethcontract::{H160, U256};
 use model::u256_decimal;
 use reqwest::{Client, IntoUrl, Url};
 use serde::Deserialize;
-use shared::solver_utils::{deserialize_prefixed_hex, Slippage};
 use std::fmt::{self, Display, Formatter};
 
 /// Parts to split a swap.
@@ -122,7 +120,7 @@ impl SwapQuery {
 }
 
 /// A 1Inch API swap response.
-#[derive(Clone, Debug, Deserialize, PartialEq, From)]
+#[derive(Clone, Debug, Deserialize, PartialEq)]
 #[serde(untagged)]
 pub enum SwapResponse {
     Swap(Box<Swap>),
@@ -134,15 +132,6 @@ pub enum SwapResponse {
 pub struct SwapResponseError {
     pub status_code: u32,
     pub description: String,
-}
-
-impl From<SwapResponseError> for SettlementError {
-    fn from(error: SwapResponseError) -> Self {
-        SettlementError {
-            inner: anyhow!(error.description),
-            retryable: matches!(error.status_code, 500),
-        }
-    }
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Default)]
@@ -217,7 +206,7 @@ pub struct Protocols {
 }
 
 // Mockable version of API Client
-#[cfg_attr(test, mockall::automock)]
+#[mockall::automock]
 #[async_trait::async_trait]
 pub trait OneInchClient: Send + Sync {
     /// Retrieves a swap for the specified parameters from the 1Inch API.
@@ -285,6 +274,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::addr;
 
     #[test]
     fn slippage_from_basis_points() {
@@ -312,10 +302,10 @@ mod tests {
     fn swap_query_serialization() {
         let base_url = Url::parse("https://api.1inch.exchange/").unwrap();
         let url = SwapQuery {
-            from_token_address: shared::addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
-            to_token_address: shared::addr!("111111111117dc0aa78b770fa6a738034120c302"),
+            from_token_address: addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
+            to_token_address: addr!("111111111117dc0aa78b770fa6a738034120c302"),
             amount: 1_000_000_000_000_000_000u128.into(),
-            from_address: shared::addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
+            from_address: addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
             slippage: Slippage::percentage_from_basis_points(50).unwrap(),
             protocols: None,
             disable_estimate: None,
@@ -341,10 +331,10 @@ mod tests {
     fn swap_query_serialization_options_parameters() {
         let base_url = Url::parse("https://api.1inch.exchange/").unwrap();
         let url = SwapQuery {
-            from_token_address: shared::addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
-            to_token_address: shared::addr!("111111111117dc0aa78b770fa6a738034120c302"),
+            from_token_address: addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
+            to_token_address: addr!("111111111117dc0aa78b770fa6a738034120c302"),
             amount: 1_000_000_000_000_000_000u128.into(),
-            from_address: shared::addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
+            from_address: addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
             slippage: Slippage::percentage_from_basis_points(50).unwrap(),
             protocols: Some(vec!["WETH".to_string(), "UNISWAP_V3".to_string()]),
             disable_estimate: Some(true),
@@ -428,10 +418,10 @@ mod tests {
             swap,
             SwapResponse::Swap(Box::new(Swap {
                 from_token: Token {
-                    address: shared::addr!("eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee"),
+                    address: addr!("eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee"),
                 },
                 to_token: Token {
-                    address: shared::addr!("111111111117dc0aa78b770fa6a738034120c302"),
+                    address: addr!("111111111117dc0aa78b770fa6a738034120c302"),
                 },
                 from_token_amount: 1_000_000_000_000_000_000u128.into(),
                 to_token_amount: 501_739_725_821_378_713_485u128.into(),
@@ -439,23 +429,19 @@ mod tests {
                     vec![Protocol {
                         name: "WETH".to_owned(),
                         part: 100.,
-                        from_token_address: shared::addr!(
-                            "eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee"
-                        ),
+                        from_token_address: addr!("eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee"),
                         to_token_address: testlib::tokens::WETH,
                     }],
                     vec![Protocol {
                         name: "UNISWAP_V2".to_owned(),
                         part: 100.,
-                        from_token_address: shared::addr!(
-                            "c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
-                        ),
-                        to_token_address: shared::addr!("111111111117dc0aa78b770fa6a738034120c302"),
+                        from_token_address: addr!("c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"),
+                        to_token_address: addr!("111111111117dc0aa78b770fa6a738034120c302"),
                     }],
                 ]],
                 tx: Transaction {
-                    from: shared::addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
-                    to: shared::addr!("11111112542d85b3ef69ae05771c2dccff4faa26"),
+                    from: addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
+                    to: addr!("11111112542d85b3ef69ae05771c2dccff4faa26"),
                     data: hex::decode(
                         "2e95b6c8\
                          0000000000000000000000000000000000000000000000000000000000000000\
@@ -502,7 +488,7 @@ mod tests {
         assert_eq!(
             spender,
             Spender {
-                address: shared::addr!("11111112542d85b3ef69ae05771c2dccff4faa26"),
+                address: addr!("11111112542d85b3ef69ae05771c2dccff4faa26"),
             }
         )
     }
@@ -513,10 +499,10 @@ mod tests {
         let swap = OneInchClientImpl::new(OneInchClientImpl::DEFAULT_URL, Client::new())
             .unwrap()
             .get_swap(SwapQuery {
-                from_token_address: shared::addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
-                to_token_address: shared::addr!("111111111117dc0aa78b770fa6a738034120c302"),
+                from_token_address: addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
+                to_token_address: addr!("111111111117dc0aa78b770fa6a738034120c302"),
                 amount: 1_000_000_000_000_000_000u128.into(),
-                from_address: shared::addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
+                from_address: addr!("00000000219ab540356cBB839Cbe05303d7705Fa"),
                 slippage: Slippage::percentage_from_basis_points(50).unwrap(),
                 protocols: None,
                 disable_estimate: None,
@@ -536,10 +522,10 @@ mod tests {
         let swap = OneInchClientImpl::new(OneInchClientImpl::DEFAULT_URL, Client::new())
             .unwrap()
             .get_swap(SwapQuery {
-                from_token_address: shared::addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
-                to_token_address: shared::addr!("a3BeD4E1c75D00fa6f4E5E6922DB7261B5E9AcD2"),
+                from_token_address: addr!("EeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"),
+                to_token_address: addr!("a3BeD4E1c75D00fa6f4E5E6922DB7261B5E9AcD2"),
                 amount: 100_000_000_000_000_000_000u128.into(),
-                from_address: shared::addr!("4e608b7da83f8e9213f554bdaa77c72e125529d0"),
+                from_address: addr!("4e608b7da83f8e9213f554bdaa77c72e125529d0"),
                 slippage: Slippage::percentage_from_basis_points(50).unwrap(),
                 protocols: Some(vec!["WETH".to_string(), "UNISWAP_V2".to_string()]),
                 disable_estimate: Some(true),
